# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u5947\u8d27API\u6587\u6863",
        "description": "\u9879\u76ee\u4ee3\u53f7Tiny(\u5c0f\u5c0f)",
        "termsOfService": "https://github.com/powrlaw",
        "contact": {
            "name": "xieyi",
            "url": "http://github.com/xieyi64",
            "email": "dev001@powerlaw.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "xieyi.api.vqiho.com",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/xml",
        "application/json"
    ],
    "produces": [
        "application/xml",
        "application/json"
    ],
    "paths": {
        "/auth/id": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6743\u9650"
                ],
                "summary": "\u5f97\u5230ID",
                "description": "\u6682\u65e0",
                "operationId": "auth.id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "\u8bbe\u5907",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "device_info",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "\u5ba2\u6237\u7aef\u7c7b\u578b",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/auth/bootstrap": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6743\u9650"
                ],
                "summary": "\u521d\u59cb\u5316",
                "description": "\u6682\u65e0",
                "operationId": "auth.bootstrap",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "\u8bbe\u5907",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "device_info",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": "{\"os\":10000}"
                    },
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "\u5ba2\u6237\u7aef\u7c7b\u578b",
                        "type": "string",
                        "default": "20100"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/auth/login": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6743\u9650"
                ],
                "summary": "\u767b\u5f55",
                "description": "\u6682\u65e0",
                "operationId": "auth.login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "loginame",
                        "in": "query",
                        "description": "\u7528\u6237\u540d|\u90ae\u7bb1|\u624b\u673a\u53f7\u7801|\u7528\u6237ID",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u5bc6\u7801",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "\u8bbe\u5907",
                        "type": "string",
                        "default": "c6c34af361e26e3967531ff8b851d674d7f79b72"
                    },
                    {
                        "name": "device_info",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": "{\"os\":10000}"
                    },
                    {
                        "name": "push_token",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": "43c23864b7dc121381cf7d71b556a03128258ae5af73d3f184e3ec5f503a628f"
                    },
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "\u5ba2\u6237\u7aef\u7c7b\u578b",
                        "type": "string",
                        "default": "20100"
                    },
                    {
                        "name": "sns_id",
                        "in": "query",
                        "description": "\u793e\u4ea4\u7f51\u7edcID",
                        "type": "string",
                        "default": "1714926623"
                    },
                    {
                        "name": "union_id",
                        "in": "query",
                        "description": "\u5fae\u4fe1\u5f00\u653e\u5e73\u53f0 Union ID",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "sns_type",
                        "in": "query",
                        "description": "\u793e\u4ea4\u7f51\u7edc\u7c7b\u578b",
                        "type": "string",
                        "default": "200"
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "\u6635\u79f0",
                        "type": "string",
                        "default": "nickname"
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "\u5934\u50cf",
                        "type": "string",
                        "default": "http://tp4.sinaimg.cn/1714926623/180/5653481386/1"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/auth/register": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6743\u9650"
                ],
                "summary": "\u6ce8\u518c",
                "description": "\u6682\u65e0",
                "operationId": "auth.register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "loginame",
                        "in": "query",
                        "description": "\u7528\u6237\u540d|\u90ae\u7bb1|\u624b\u673a\u53f7\u7801|\u7528\u6237ID",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u5bc6\u7801",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "\u8bbe\u5907",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "device_info",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": "{\"os\":10000}"
                    },
                    {
                        "name": "push_token",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "\u5ba2\u6237\u7aef\u7c7b\u578b",
                        "type": "string",
                        "default": "20100"
                    },
                    {
                        "name": "sns_id",
                        "in": "query",
                        "description": "\u793e\u4ea4\u7f51\u7edcID",
                        "type": "string",
                        "default": "1714926623"
                    },
                    {
                        "name": "union_id",
                        "in": "query",
                        "description": "\u5fae\u4fe1\u5f00\u653e\u5e73\u53f0 Union ID",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "sns_type",
                        "in": "query",
                        "description": "\u793e\u4ea4\u7f51\u7edc\u7c7b\u578b",
                        "type": "string",
                        "default": "200"
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "\u6635\u79f0",
                        "type": "string",
                        "default": "nickname"
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "\u5934\u50cf",
                        "type": "string",
                        "default": "http://tp4.sinaimg.cn/1714926623/180/5653481386/1"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6743\u9650"
                ],
                "summary": "\u9000\u51fa",
                "description": "\u6682\u65e0",
                "operationId": "auth.logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/auth/bind": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6743\u9650"
                ],
                "summary": "\u7ed1\u5b9a",
                "description": "\u6682\u65e0",
                "operationId": "auth.bind",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "\u7528\u6237ID",
                        "type": "string",
                        "default": "100000000000000000"
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "\u8bbe\u5907",
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "device_info",
                        "in": "query",
                        "description": "\u8bbe\u5907\u4fe1\u606f",
                        "type": "string",
                        "default": "{\"os\":10000}"
                    },
                    {
                        "name": "client_type",
                        "in": "query",
                        "description": "\u5ba2\u6237\u7aef\u7c7b\u578b",
                        "type": "string",
                        "default": "20100"
                    },
                    {
                        "name": "sns_id",
                        "in": "query",
                        "description": "\u793e\u4ea4\u7f51\u7edcID",
                        "type": "string",
                        "default": "1714926623"
                    },
                    {
                        "name": "sns_type",
                        "in": "query",
                        "description": "\u793e\u4ea4\u7f51\u7edc\u7c7b\u578b",
                        "type": "string",
                        "default": "200"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "\u6307\u5b9a\u5b57\u6bb5(id,name)",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
...
}